Travis now supports R as a top-level language
I used the website for R-packages to submit my first package to CRAN. Thanks for a great resource!! I noticed that this section was out of date. My suggestions are in the proposed change. 

From the official documentation: "The R environment comes with LaTeX and pandoc preinstalled, making it easier to use packages like RMarkdown or knitr." This suggests BOOTSTRAP_LATEX=1 is no longer needed.

I didn't see any mention of "`_R_CHECK_FORCE_SUGGESTS_=0`" so I am not sure if it is still available as an option.