Added text about Rtools paths
I was receiving error 127 when I ran has_devel().  After a lot of sleuthing, I found that the default setting when installing Rtools did not update the system path.  I have added language to the instructions for Rtools installation that warns the user to check the box to allow the installation package to update the system path.

“I assign the copyright of this contribution to Hadley Wickham