`Imports` assume installing CRAN version
I found this in a really hard way. 

I tested my package a lot, put it to github, tried to install it to my laptop, then there were lots of errors. At first the error seemed to be `data.table` awareness problem, but further debugging point to a Shiny DT rowselected value to be character in laptop. Turned out I installed the newest github version `DT` in my PC since this is recommended by `DT` website, but my laptop only installed the CRAN version of `DT` because the Imports field only assume this.