Full stops additions & deletions in documentation (#548)
The guidelines say that titles should not end in a full stop, and parameter descriptions should end in a full stop. Several examples in this chapter
violated against this rule, some fixes are proposed. It remains to be further evaluated by the authors whether this must also hold for object
documentation (this commit did not fix that, however the book is unclear about that).